plugin-base: improve the logic to check whether vendor or product probing needed
If the plugin does a Vendor ID check and it passes, it doesn't need Vendor
string probing.

If the plugin does a Vendor ID check and it fails:
  - If Vendor strings reported, it needs Vendor probing.
  - If Vendor strings not reported, fail as unsupported.

If the plugin does a Product ID check and it passes, it doesn't need Product
string probing.

If the plugin does a Product ID check and it fails:
  - If Product strings reported, it needs Product probing.
  - If Product strings not reported, fail as unsupported.