plugin-base: avoid trying to probe 'net' devices
Some devices support 'net' devices, which cannot be AT or QCDM probed.

If the port being checked for support corresponds to an already existing modem,
the port will be reported as SUPPORTED.

If this is the first port of the modem being checked for support, we will just
DEFER the support check until we get a modem created.