broadband-modem: try to simplify 3GPP network registration
When requesting to register in the network manually, we will poll every once
in a while to check whether we got registered or not. We remove the registration
timeout and we handle the case where it never gets registered by allowing a
maximum number of registration checks in our request.