iface-modem-cdma: allow to configure which registration steps to run
Implementations of the CDMA interface are allowed to specify which steps in the
registration checks can be skipped. For example, if no QCDM ports are available,
there won't be any need to try QCDM-based checks; or if Sprint commands are
available, there won't be any need to run CDMA1x Serving System checks.