broadband-modem-qmi: enable/disable signal quality related unsolicited events
The logic here involves knowing if we have support for 'Signal Info'
indications. All the 'Signal Info' stuff was included in NAS 1.8, so we're
just using a single flag to indicate support for that.

If we do have 'Signal Info', enabling the signal quality indications involves
two steps: configuring thresholds with "Config Signal Info" and then enabling
them with "Register Indications".

If we don't have 'Signal Info' we fallback to the good old "Set Event Report"
for enabling/disabling the indications.