broadband-modem-qmi: use only deprecated commands
The QMI protocol is updated frequently with new commands replacing and
improving some already existing ones, which get marked as deprecated.
Instead of initially try to cover the logic with both the old and the new
commands, we'll default to use only the old deprecated ones, which should be
supported also in the modems with newer services.

The supported already implemented for the new QMI commands is hidden behind a
`--with-newest-qmi-commands' configure switch.

Once the implementation is fully developed and tested using the deprecated
commands we can then focus in including the support for the new ones.