huawei: better detection of data port on some modems (lp:1057186)
Some devices (E173) appear to lie about NDIS support; GETPORTMODE
reports NDIS is enabled, but that port is actually the MDM port and
responds to AT commands.  So if we have an NDIS port but no MDM port,
and SETPORT or U2DIAG report NDIS is disabled, assume the NDIS port
is the MDM port.