via: plugin has vendor/product string probing, so needs to be sorted last
The flag to get the plugin sorted last allows us to ensure that all plugins
without vendor/product string probing (e.g. vendor/product ID probing) are run
before. We don't want the Via plugin to start probing e.g. Huawei modems.

In git master this flag in the plugin is not needed, it is automatically ordered
last if vendor/product ID probing is set in the plugin.