sierra: cache SIM PIN and re-send it after power-up if needed
When we recover from low-power mode without a modem reset, we may need to
re-send the PIN:

 (ttyUSB6): --> 'AT+CFUN?<CR>'
 (ttyUSB8): <-- '<CR><LF>OK<CR><LF>'
 (ttyUSB6): <-- '<CR><LF>+CFUN: 4<CR><LF><CR><LF>OK<CR><LF>'
 (ttyUSB6): --> 'AT+CFUN=1,0<CR>'
 (ttyUSB6): <-- '<CR><LF>OK<CR><LF>'
 (ttyUSB6): --> 'AT+CPIN?<CR>'
 (ttyUSB6): <-- '<CR><LF>+CPIN: SIM PIN<CR><LF><CR><LF>OK<CR><LF>'
 (ttyUSB6): --> 'AT+CPIN="1234"<CR>'
 (ttyUSB6): <-- '<CR><LF>OK<CR><LF>'

The cached SIM PIN will be the last SIM PIN used to unlock the modem.

If, for any reason, sending the cached PIN fails, we'll end up removing the
cache, so that the user doesn't try to enable the modem too many times using
a possible wrong PIN.

If we get PIN-locked after the power-up and we didn't cache any PIN, which may
happen if MM crashes, reboots and the modem was already unlocked (we won't ask
for PIN during initialization), then we'll perform a device reset. This
situation is really a corner case to be handled nicely.